rghhth

+++ src/Modules/shop/controller/deleteShopControler.js
@@ -2,6 +2,10 @@ const Shop = require("../model/shopSchema.js");
 
 const deleteShop = async (req, res, next) => {
   try {
+    const id = req.body.vendorId;
+    if (!id) throw new Error("VendorId is required");
+    const data = await Shop.findById({ _id: id });
+    deleteImage(data.image);
     const shop = await Shop.findOneAndDelete(req.vendorId, { isActive: false });
 
     return res.status(200).json({
